requirements.txt
bcrypt==4.2.0
google-cloud-storage==2.18.2
pandas==2.2.2
streamlit==1.38.0
requests==2.32.3
google-cloud-secret-manager==2.20.2
google-cloud-logging==3.11.2